# Lazily initialize RangeDelAggregator's map and pinning manager